Add tokenExpiryTime option (#74)
* Add tokenExpiryTime option

This allows the user to override the time it takes for the api token
to expire. It defaults to 5 minutes to prevent making this a breaking
change. The API will most likely start using 20 minutes as the default
length in a future update (https://github.com/directus/api/pull/913).
Once that update is through, we can change the 5 min default in the
sdk to 20 too.

* Use local token exp in refresh